Mejorar sistema de UnitTests para controladores y dao evitando Knex y sqlite.
Squashed commit of the following:

commit 5c2273fb9ccf090aae6028cb1cd89fac0c90dd38
Author: iago sardiña <iago.sardina@landrasistemas.es>
Date:   Thu Dec 16 15:15:09 2021 +0100

    Improve fakeResponse

commit ff7433c7aa66102056d5f93eb773535d915b6ea7
Author: iago sardiña <iago.sardina@landrasistemas.es>
Date:   Thu Dec 16 15:10:40 2021 +0100

    Fix test command.

commit 3967fd6ce03bc664dcd846d42f526bb4981f6dc6
Author: iago sardiña <iago.sardina@landrasistemas.es>
Date:   Thu Dec 16 13:48:20 2021 +0100

    Mejorar unit testing. Mocks para la base de datos y otros metodos de App.

commit c7ed4d7363e307e92f035a4536a9494fbb8ff64d
Author: iago sardiña <iago.sardina@landrasistemas.es>
Date:   Thu Dec 16 08:18:30 2021 +0100

commit 80d14137eefbef4bdc641593db024934b0012571
Author: iago sardiña <iago.sardina@landrasistemas.es>
Date:   Tue Dec 14 16:48:31 2021 +0100

commit 976b5bf4b0496a9c93a44516a6559d3d38718e4e
Merge: bde8967 beeda0b
Author: iago sardiña <iago.sardina@landrasistemas.es>
Date:   Tue Dec 14 15:24:02 2021 +0100

    Merge branch 'main' into copy-paste

commit bde8967472ddf1b22311b8742f93688068c857c5
Author: iago sardiña <iago.sardina@landrasistemas.es>
Date:   Tue Dec 14 13:10:54 2021 +0100

+++ app/server/integration/jum-angie/JumDao.js
@@ -7,6 +7,9 @@ export default class JumDao {
     }
 
     async deployRoute(id, content, options) {
+        if (!this.jum_url) {
+            return null;
+        }
         const response = await axios.post(this.jum_url + "/create", {
             id: id,
             xmlContent: content,
@@ -21,6 +24,9 @@ export default class JumDao {
     }
 
     async undeployRoute(id) {
+        if (!this.jum_url) {
+            return null;
+        }
         const response = await axios.delete(this.jum_url + "/undeploy/" + id);
 
         if (response.status != 200) {
@@ -30,6 +36,9 @@ export default class JumDao {
     }
 
     async startRoute(id) {
+        if (!this.jum_url) {
+            return null;
+        }
         const response = await axios.put(this.jum_url + "/start/" + id);
 
         if (response.status != 200) {
@@ -39,6 +48,9 @@ export default class JumDao {
     }
 
     async stopRoute(id) {
+        if (!this.jum_url) {
+            return null;
+        }
         const response = await axios.put(this.jum_url + "/stop/" + id);
 
         if (response.status != 200) {
@@ -48,6 +60,9 @@ export default class JumDao {
     }
 
     async getRouteStatus(id) {
+        if (!this.jum_url) {
+            return null;
+        }
         console.log("obteniendo estado canal");
         const response = await axios.get(this.jum_url + "/get/" + id);
 
@@ -58,6 +73,9 @@ export default class JumDao {
     }
 
     async list() {
+        if (!this.jum_url) {
+            return null;
+        }
         const response = await axios.get(this.jum_url + "/list");
 
         if (response.status != 200) {
@@ -67,6 +85,9 @@ export default class JumDao {
     }
 
     async getRouteStats(id) {
+        if (!this.jum_url) {
+            return null;
+        }
         const response = await axios.get(this.jum_url + "/get/" + id + "?includeStats=true");
 
         if (response.status != 200) {
@@ -76,6 +97,9 @@ export default class JumDao {
     }
 
     async getRouteLogs(id) {
+        if (!this.jum_url) {
+            return null;
+        }
         const response = await axios.get(this.jum_url + "/log/" + id);
 
         if (response.status != 200) {
@@ -85,6 +109,9 @@ export default class JumDao {
     }
 
     async sendMessageToRoute(id, endpoint, content) {
+        if (!this.jum_url) {
+            return null;
+        }
         const response = await axios.post(this.jum_url + "/sendMessageToRoute/" + id, {
             endpoint: endpoint,
             content: content,
